refactor(frontend): split playtest marker runtime
