+++ engine/examples/editor-playtest/src/marker_runtime.rs
@@ -1,0 +1,149 @@
+use super::*;
+
+const TARGET_LOCK_OUTER: i32 = 25;
+const TARGET_LOCK_INNER: i32 = 13;
+const TARGET_LOCK_TRI_HALF_WIDTH: i32 = 8;
+const TARGET_LOCK_RED: (u8, u8, u8) = (225, 18, 24);
+const TARGET_LOCK_ROTATION_FRAMES: u32 = 360;
+
+/// Marker visualization tuning. Markers are debug stubs -- keep
+/// them visible at orbit-camera scales without dominating the
+/// scene.
+const MARKER_HALF: i32 = 96;
+const MARKER_LIFT: i32 = MARKER_HALF;
+const MARKER_TINT: (u8, u8, u8) = (0xff, 0xa8, 0x40);
+
+/// Draw one tinted cube per generated entity record. Cubes
+/// reuse the room's first material with an override tint so
+/// markers stand out from the surrounding geometry without
+/// needing a dedicated texture upload.
+pub(super) fn draw_entity_markers(
+    entities: &[EntityRecord],
+    current_room: RoomIndex,
+    materials: &[WorldRenderMaterial],
+    camera: &WorldCamera,
+    options: WorldSurfaceOptions,
+    triangles: &mut impl PrimitiveSink<TriTextured>,
+    world: &mut WorldRenderPass<'_, '_, OT_DEPTH>,
+) {
+    if entities.is_empty() || materials.is_empty() {
+        return;
+    }
+    // Reuse the room's first material so we don't need a
+    // dedicated marker texture. Tint override picks up the
+    // existing CLUT + tpage but recolours.
+    let material = materials[0].texture.with_tint(MARKER_TINT);
+    let opts = options.with_material_layer(material);
+    const UVS: [(u8, u8); 4] = [(0, 0), (64, 0), (64, 64), (0, 64)];
+
+    for entity in entities {
+        if entity.room != current_room {
+            continue;
+        }
+        let cx = entity.x;
+        let cy = entity.y - MARKER_LIFT - MARKER_HALF;
+        let cz = entity.z;
+        let h = MARKER_HALF;
+
+        let top = [
+            WorldVertex::new(cx - h, cy - h, cz - h),
+            WorldVertex::new(cx + h, cy - h, cz - h),
+            WorldVertex::new(cx + h, cy - h, cz + h),
+            WorldVertex::new(cx - h, cy - h, cz + h),
+        ];
+        let bottom = [
+            WorldVertex::new(cx - h, cy + h, cz + h),
+            WorldVertex::new(cx + h, cy + h, cz + h),
+            WorldVertex::new(cx + h, cy + h, cz - h),
+            WorldVertex::new(cx - h, cy + h, cz - h),
+        ];
+        let north = [
+            WorldVertex::new(cx - h, cy - h, cz - h),
+            WorldVertex::new(cx + h, cy - h, cz - h),
+            WorldVertex::new(cx + h, cy + h, cz - h),
+            WorldVertex::new(cx - h, cy + h, cz - h),
+        ];
+        let south = [
+            WorldVertex::new(cx + h, cy - h, cz + h),
+            WorldVertex::new(cx - h, cy - h, cz + h),
+            WorldVertex::new(cx - h, cy + h, cz + h),
+            WorldVertex::new(cx + h, cy + h, cz + h),
+        ];
+        let east = [
+            WorldVertex::new(cx + h, cy - h, cz - h),
+            WorldVertex::new(cx + h, cy - h, cz + h),
+            WorldVertex::new(cx + h, cy + h, cz + h),
+            WorldVertex::new(cx + h, cy + h, cz - h),
+        ];
+        let west = [
+            WorldVertex::new(cx - h, cy - h, cz + h),
+            WorldVertex::new(cx - h, cy - h, cz - h),
+            WorldVertex::new(cx - h, cy + h, cz - h),
+            WorldVertex::new(cx - h, cy + h, cz + h),
+        ];
+
+        for face in [top, bottom, north, south, east, west] {
+            if let Some(projected) = camera.project_world_quad(face) {
+                let _ = world.submit_textured_quad(triangles, projected, UVS, material, opts);
+            }
+        }
+    }
+}
+
+pub(super) fn draw_lock_target_indicator(
+    target: RoomPoint,
+    camera: WorldCamera,
+    elapsed_tick: SimTick,
+) {
+    let Some(center) = camera.project_world(target) else {
+        return;
+    };
+
+    let outer = TARGET_LOCK_OUTER;
+    let inner = TARGET_LOCK_INNER;
+    let half_width = TARGET_LOCK_TRI_HALF_WIDTH;
+    let angle = Angle::per_frames(TARGET_LOCK_ROTATION_FRAMES).mul_tick(elapsed_tick);
+    let triangles = [
+        [
+            target_screen_vertex(center, 0, -inner, angle),
+            target_screen_vertex(center, -half_width, -outer, angle),
+            target_screen_vertex(center, half_width, -outer, angle),
+        ],
+        [
+            target_screen_vertex(center, 0, inner, angle),
+            target_screen_vertex(center, half_width, outer, angle),
+            target_screen_vertex(center, -half_width, outer, angle),
+        ],
+        [
+            target_screen_vertex(center, -inner, 0, angle),
+            target_screen_vertex(center, -outer, half_width, angle),
+            target_screen_vertex(center, -outer, -half_width, angle),
+        ],
+        [
+            target_screen_vertex(center, inner, 0, angle),
+            target_screen_vertex(center, outer, -half_width, angle),
+            target_screen_vertex(center, outer, half_width, angle),
+        ],
+    ];
+
+    for triangle in triangles {
+        draw_tri_flat_blended(
+            triangle,
+            TARGET_LOCK_RED.0,
+            TARGET_LOCK_RED.1,
+            TARGET_LOCK_RED.2,
+            BlendMode::Average,
+        );
+    }
+}
+
+fn target_screen_vertex(center: ProjectedVertex, ox: i32, oy: i32, angle: Angle) -> (i16, i16) {
+    let sin = angle.sin_q12();
+    let cos = angle.cos_q12();
+    let rx = ((ox.saturating_mul(cos)).saturating_sub(oy.saturating_mul(sin))) >> 12;
+    let ry = ((ox.saturating_mul(sin)).saturating_add(oy.saturating_mul(cos))) >> 12;
+    (
+        clamp_i16((center.sx as i32).saturating_add(rx)),
+        clamp_i16((center.sy as i32).saturating_add(ry)),
+    )
+}
